pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "mcp>=1.9.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
